Add `Deprecations+Removals.swift` and move `AWSAppSyncClientError` deprecations into
